chore: move Prefix to a new file
