+++ crates/rattler_conda_types/src/lib.rs
@@ -24,6 +24,7 @@ mod generic_virtual_package;
 pub mod package;
 mod package_name;
 pub mod prefix_record;
+pub mod prefix;
 
 #[cfg(test)]
 use std::path::{Path, PathBuf};
